+++ src/main/kotlin/com/flab/ticketing/order/repository/reader/CartReader.kt
@@ -1,0 +1,3 @@
+package com.flab.ticketing.order.repository.reader
+
+class CartReader
