--- src/main/kotlin/com/flab/ticketing/order/service/reader/CartReader.kt
@@ -1,4 +1,0 @@
-package com.flab.ticketing.order.service.reader
-
-class CartReader {
-}
